app/requirements.txt
apispec==3.3.0
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
connect-mysql==0.0.5
cryptography==2.9.2
Flask==1.1.1
flask-smorest==0.18.5
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.6.1
mysql-connector==2.2.9
mysql-connector-python==8.0.20
nested-lookup==0.2.21
numpy==1.18.5
pandas==0.24.1
paramiko==2.7.1
protobuf==3.12.2
pycparser==2.20
PyMySQL==0.9.3
PyNaCl==1.4.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.22.0
six==1.15.0
sshtunnel==0.1.5
urllib3==1.25.9
webargs==5.5.3
Werkzeug==1.0.1
xmltodict==0.12.0
